+++ src/SurviveTheHuntClient/MainScript.cs
@@ -19,7 +19,7 @@ using SurviveTheHuntShared;
 
 namespace SurviveTheHuntClient
 {
-    public class MainScript : ClientScript
+    public partial class MainScript : ClientScript
     {
         /// <summary>
         /// Local player state maintained by the client script.
@@ -59,6 +59,33 @@ namespace SurviveTheHuntClient
         /// Blips used to represent player deaths on the radar.
         /// </summary>
         private readonly DeathBlips DeathBlips;
+        
+        /// <summary>
+        /// The local player's ped handle as of the previous tick. Used to track ped changes so max health can be applied correctly.
+        /// </summary>
+        private int PreviousTickPedHandle = 0;
+
+        /// <summary>
+        /// Was prep phase active during the last tick?
+        /// </summary>
+        private bool DidProtectionsApplyLastFrame = false;
+
+        /// <summary>
+        /// Handle to the blip displaying the safe zone radius.
+        /// </summary>
+        private int SafeZoneRadiusBlipHandle = 0;
+
+        /// <summary>
+        /// Has the player already spawned at least once?
+        /// </summary>
+        private bool SpawnedOnce = false;
+
+        /// <summary>
+        /// Clipset to apply to the player ped.
+        /// </summary>
+        private string TargetClipset = "";
+
+        private bool IsTargetClipsetLoaded = false;
 
         public MainScript()
         {
@@ -81,6 +108,11 @@ namespace SurviveTheHuntClient
             if(GetCurrentResourceName() != resourceName)
             {
                 return;
+            }
+
+            if(DoesBlipExist(SafeZoneRadiusBlipHandle))
+            {
+                RemoveBlip(ref SafeZoneRadiusBlipHandle);
             }
 
             foreach (SyncedVehicle vehicleToDelete in SpawnedVehicles)
@@ -152,6 +184,11 @@ namespace SurviveTheHuntClient
 
                 // Notify the server this client has started so the config can be sent down. This is needed for resource restarts etc.
                 TriggerServerEvent(Events.Server.ClientStarted);
+
+                SafeZoneRadiusBlipHandle = AddBlipForRadius(SharedConstants.DockSpawn.X, SharedConstants.DockSpawn.Y, SharedConstants.DockSpawn.Z, SharedConstants.DefaultSpawnSafeZoneRadius);
+                SetBlipColour(SafeZoneRadiusBlipHandle, 69);
+                SetBlipAlpha(SafeZoneRadiusBlipHandle, 128);
+                SetBlipDisplay(SafeZoneRadiusBlipHandle, 6);
             }
         }
 
@@ -254,11 +291,24 @@ namespace SurviveTheHuntClient
         {
             Vector3 spawnLoc = SharedConstants.DockSpawn;
 
-            Exports["spawnmanager"].spawnPlayer(new { x = spawnLoc.X, y = spawnLoc.Y, z = spawnLoc.Z, model = "a_m_m_skater_01" });
+            int randomPedIndex = RNG.Next(0, SharedConstants.DefaultPlayerPeds.Length);
+            string randomPed = SharedConstants.DefaultPlayerPeds[randomPedIndex];
+
+            Exports["spawnmanager"].spawnPlayer(new { x = spawnLoc.X, y = spawnLoc.Y, z = spawnLoc.Z, model = randomPed });
+
+            // Pick a random GTAO character clipset.
+            bool isFemale = randomPed.StartsWith("a_f");
+            string[] clipsets = isFemale ? SharedConstants.DefaultFemaleClipsets : SharedConstants.DefaultMaleClipsets;
+            TargetClipset = clipsets[RNG.Next(0, clipsets.Length)];
+            IsTargetClipsetLoaded = false;
+
+            DidProtectionsApplyLastFrame = false;
         }
 
         protected void PlayerSpawnedCallback()
         {
+            SpawnedOnce = true;
+
             // Refresh player's death state.
             PlayerState.DeathReported = false;
 
@@ -277,6 +327,30 @@ namespace SurviveTheHuntClient
                 Debug.WriteLine($"GameState: Hunt.IsInProgress: {GameState.Hunt.IsInProgress}, Hunt.IsEnding: {GameState.Hunt.IsEnding}");
                 HuntUI.DisplayObjective(ref GameState, ref PlayerState, GameState.Hunt.IsEnding);
             }
+
+            // Set the player's max health.
+            ApplyMaxHealth(true);
+
+            if (SharedConstants.DefaultPlayerPeds.Any(modelName => GetHashKey(modelName) == Player.Local.Character.Model.Hash))
+            {
+                SetPedRandomComponentVariation(Player.Local.Character.Handle, false);
+                SetPedRandomProps(Player.Local.Character.Handle);
+            }
+        }
+
+        /// <summary>
+        /// Applies max health to the current player ped and optionally replenishes their health.
+        /// </summary>
+        /// <param name="restore">Should the player ped's health be replenished to max?</param>
+        protected void ApplyMaxHealth(bool restore = false)
+        {
+            int maxHealth = GetConvarInt("sth_maxHealth", SharedConstants.DefaultMaxHealth);
+            SetPedMaxHealth(PlayerPedId(), maxHealth);
+            
+            if (restore)
+            {
+                SetEntityHealth(PlayerPedId(), maxHealth);
+            }
         }
 
         protected async Task UpdateLoop()
@@ -285,7 +359,23 @@ namespace SurviveTheHuntClient
             {
                 ResetPlayerStamina(PlayerId());
                 PlayerPos = Game.PlayerPed.Position;
+
+                // Wait for the random clipset to load and apply it once it's loaded.
+                bool wasClipsetLoadedLastFrame = IsTargetClipsetLoaded;
+                if(!wasClipsetLoadedLastFrame)
+                {
+                    RequestClipSet(TargetClipset);
+                }
+                IsTargetClipsetLoaded = HasClipSetLoaded(TargetClipset);
+                if (!wasClipsetLoadedLastFrame && IsTargetClipsetLoaded)
+                {
+                    SetPedMovementClipset(PlayerPedId(), TargetClipset, 1f);
+                }
             }
+
+            bool wasHuntStartedLastFrame = !GameState.Hunt.WasHuntInProgressLastFrame && GameState.Hunt.IsStarted;
+
+            GameState.Hunt.Tick();
 
             GameState.Hunt.UpdateHuntedMugshot();
             HuntUI.SetBigmap(ref PlayerState);
@@ -329,9 +419,114 @@ namespace SurviveTheHuntClient
 
             FixCarsInSpawn();
 
+            CfxVector3 spawnPos = new CfxVector3(SharedConstants.DockSpawn.X, SharedConstants.DockSpawn.Y, SharedConstants.DockSpawn.Z);
+            float safeZoneRadiusSqr = (SharedConstants.DefaultSpawnSafeZoneRadius * SharedConstants.DefaultSpawnSafeZoneRadius);
+
+            // Apply spawn safe zone protections if the prep phase hasn't ended yet.
+            bool inSafeZone = Player.Local?.Character?.Position != null
+                ? Player.Local.Character.Position.DistanceToSquared(spawnPos) < safeZoneRadiusSqr
+                : false;
+            bool isPrepPhase = (GameState.Hunt.IsStarted && GameState.Hunt.IsPrepPhase);
+            bool shouldProtectionsApply = isPrepPhase || inSafeZone;
+
+            if (wasHuntStartedLastFrame && !inSafeZone)
+            {
+                PlayerState.WaitingToTeleportToSpawn = true;
+            }
+
+            bool canLeaveSpawn = (!isPrepPhase || PlayerState.Team == Teams.Team.Hunted) && !PlayerState.WaitingToTeleportToSpawn;
+
+            ApplySafeZoneProtection(shouldProtectionsApply, canLeaveSpawn, spawnPos, SharedConstants.DefaultSpawnSafeZoneRadius);
+            if(!canLeaveSpawn)
+            {
+                Vector3 spawn = SharedConstants.DockSpawn;
+                if(Player.Local?.Character?.Position != null && Player.Local.Character.Position.DistanceToSquared(spawnPos) > safeZoneRadiusSqr * 0.9f)
+                {
+                    DrawSphere(spawn.X, spawn.Y, spawn.Z, SharedConstants.DefaultSpawnSafeZoneRadius, 206, 191, 78, 0.5f);
+                }
+            }
+            if(!isPrepPhase && shouldProtectionsApply != DidProtectionsApplyLastFrame)
+            {
+                BeginTextCommandThefeedPost("STRING");
+                AddTextComponentSubstringPlayerName(inSafeZone ? "Safe zone protections active." : "Exited safe zone.");
+                EndTextCommandThefeedPostTicker(true, true);
+            }
+            DidProtectionsApplyLastFrame = shouldProtectionsApply;
+
             DeathBlips.ClearExpiredBlips();
 
+            if(Game.PlayerPed?.Exists() == true)
+            {
+                if(PreviousTickPedHandle != Game.PlayerPed.Handle)
+                {
+                    Debug.WriteLine("Ped changed, setting max health");
+                    ApplyMaxHealth();
+                }
+
+                PreviousTickPedHandle = Game.PlayerPed.Handle;
+            }
+
             Wait(0);
+        }
+
+        void ApplySafeZoneProtection(bool protectionActive, bool canLeaveSpawn, CfxVector3 safeZoneOrigin, float safeZoneRadius)
+        {
+            SetPlayerInvincible(PlayerId(), protectionActive);
+
+            // It's CRITICAL that you ensure this statement only runs if SpawnedOnce is true.
+            // Otherwise, players joining in progress will get softlocked without an error because the game will attempt to teleport them
+            // before they've fully loaded in.
+            if (SpawnedOnce && !canLeaveSpawn)
+            {
+                if (Game.PlayerPed?.Exists() == true)
+                {
+                    CfxVector3 offset = Game.PlayerPed.Position - safeZoneOrigin;
+                    float magnitudeSqr = offset.LengthSquared();
+                    float radiusSqr = safeZoneRadius * safeZoneRadius;
+                    bool isOut = magnitudeSqr > radiusSqr;
+
+                    if (isOut)
+                    {
+                        // Get a unit-length direction vector (from the spawn point to the player's current pos)
+                        CfxVector3 dir = offset;
+                        dir.Normalize();
+
+                        int entityId = Game.PlayerPed.IsInVehicle() ? Game.PlayerPed.CurrentVehicle.Handle : PlayerPedId();
+
+                        CfxVector3 velocity = GetEntityVelocity(entityId);
+
+                        // Get the relation between our current velocity and the direction vector (dot product of the two).
+                        CfxVector3 normVel = velocity;
+                        normVel.Normalize();
+                        float dot = normVel.X * dir.X + normVel.Y * dir.Y + normVel.Z * dir.Z;
+
+                        // Remove any velocity that points away from the spawn.
+                        if (dot >= 0f && velocity.LengthSquared() > 0f)
+                        {
+                            float mult = Math.Max(-dot, 0f);
+                            SetEntityVelocity(entityId, velocity.X * mult, velocity.Y * mult, velocity.Z * mult);
+                        }
+
+                        bool needsTeleport = PlayerState.WaitingToTeleportToSpawn || magnitudeSqr > radiusSqr * 1.2f;
+
+                        if (needsTeleport)
+                        {
+                            if (!IsScreenFadingOut())
+                            {
+                                DoScreenFadeOut(500);
+                            }
+
+                            if (IsScreenFadedOut())
+                            {
+                                SetEntityCoordsNoOffset(entityId, safeZoneOrigin.X, safeZoneOrigin.Y, safeZoneOrigin.Z, true, false, false);
+                                PlayerState.WaitingToTeleportToSpawn = false;
+                                DoScreenFadeIn(500);
+                                HuntUI.DisplayObjective(ref GameState, ref PlayerState);
+                            }
+                        }
+                    }
+                }
+            }
         }
 
         /// <summary>
@@ -376,6 +571,41 @@ namespace SurviveTheHuntClient
             }
         }
 
+        private void NotifyTeam(Teams.Team playerTeam, Player huntedPlayer)
+        {
+            Ped playerPed = Game.PlayerPed;
+            GameState.Hunt.IsStarted = true;
+            GameState.Hunt.HuntedPlayer = huntedPlayer;
+
+            switch (playerTeam)
+            {
+                case Teams.Team.Hunters:
+                    // TODO: Shouldn't current objective technically be player state?
+                    GameState.CurrentObjective = " is the hunted! Track them down.";
+                    PlayerState.Team = Teams.Team.Hunters;
+                    break;
+                case Teams.Team.Hunted:
+                    GameState.CurrentObjective = "Survive";
+                    PlayerState.Team = Teams.Team.Hunted;
+                    break;
+            }
+
+            PlayerState.TakeAwayWeapons(ref playerPed);
+        }
+
+        private void HuntStartedByServer(float secondsTillPing, DateTime endTime, TimeSpan? prepPhase = null)
+        {
+            if (!prepPhase.HasValue)
+            {
+                prepPhase = TimeSpan.Zero;
+            }
+
+            GameState.Hunt.NextMugshotTime = Utility.CurrentTime + TimeSpan.FromSeconds(secondsTillPing);
+            GameState.Hunt.InitialEndTime = endTime;
+            GameState.Hunt.PrepPhaseEndTime = Utility.CurrentTime + prepPhase.Value;
+            HuntUI.DisplayObjective(ref GameState, ref PlayerState);
+        }
+
         /// <summary>
         /// Populates <see cref="STHEvents"/> with gamemode-specific events used by the resource.
         /// </summary>
@@ -394,38 +624,21 @@ namespace SurviveTheHuntClient
                 {
                     Events.Client.NotifyHuntedPlayer.EventName(), new Action<dynamic>(data =>
                     {
-                        //Debug.WriteLine("I'm the hunted!");
-                        GameState.Hunt.IsStarted = true;
-                        GameState.Hunt.HuntedPlayer = Game.Player;
-
-                        GameState.CurrentObjective = "Survive";
-                        PlayerState.Team = Teams.Team.Hunted;
-
-                        Ped playerPed = Game.PlayerPed;
-                        PlayerState.TakeAwayWeapons(ref playerPed);
+                        NotifyTeam(Teams.Team.Hunted, Game.Player);
                     })
                 },
                 {
                     Events.Client.NotifyHunters.EventName(), new Action<dynamic>(data =>
                     {
-                        string huntedPlayerName = data.HuntedPlayerName;
+                        int huntedPlayerServerId = data.HuntedPlayerServerId;
 
                         // Since the event is sent out to everyone, make sure it is discarded by the hunted player.
-                        if(huntedPlayerName == Game.Player.Name)
+                        if(huntedPlayerServerId == Game.Player.ServerId)
                         {
                             return;
                         }
 
-                        Ped playerPed = Game.PlayerPed;
-
-                        GameState.Hunt.IsStarted = true;
-                        GameState.Hunt.HuntedPlayer = Players[huntedPlayerName];
-
-                        // TODO: Shouldn't current objective technically be player state?
-                        GameState.CurrentObjective = " is the hunted! Track them down.";
-                        PlayerState.Team = Teams.Team.Hunters;
-
-                        PlayerState.TakeAwayWeapons(ref playerPed);
+                        NotifyTeam(Teams.Team.Hunters, new Player(GetPlayerFromServerId(huntedPlayerServerId)));
                     })
                 },
                 {
@@ -449,21 +662,21 @@ namespace SurviveTheHuntClient
                     Events.Client.HuntStartedByServer.EventName(), new Action<dynamic>(data =>
                     {
                         float secondsTillPing = data.NextNotification;
-                        GameState.Hunt.NextMugshotTime = Utility.CurrentTime + TimeSpan.FromSeconds(secondsTillPing);
 
                         string endTimeStr = data.EndTime;
                         DateTime endTime = DateTime.ParseExact(endTimeStr, "F", CultureInfo.InvariantCulture);
-                        GameState.Hunt.InitialEndTime = endTime;
-                        HuntUI.DisplayObjective(ref GameState, ref PlayerState);
-                        Ped playerPed = Game.PlayerPed;
+
+                        ulong prepPhaseDuration = (ulong)data.PrepPhaseDuration;
+
+                        HuntStartedByServer(secondsTillPing, endTime, TimeSpan.FromSeconds(prepPhaseDuration));
                     })
                 },
                 {
                     Events.Client.ShowPingOnMap.EventName(), new Action<dynamic>(data =>
                     {
-                        string playerName = data.PlayerName;
-                        HuntUI.CreateRadiusBlipForPlayer(Players[playerName], data.Radius, data.OffsetX, data.OffsetY, DateTime.ParseExact(data.CreationDate, "F", CultureInfo.InvariantCulture), ref PlayerState);
-                        if(playerName == Game.Player.Name)
+                        int playerServerId = int.Parse(data.PlayerServerId);
+                        HuntUI.CreateRadiusBlipForPlayer(new Player(GetPlayerFromServerId(playerServerId)), data.Radius, data.OffsetX, data.OffsetY, DateTime.ParseExact(data.CreationDate, "F", CultureInfo.InvariantCulture), ref PlayerState);
+                        if(playerServerId == Game.Player.ServerId)
                         {
                             CfxVector3 position = GetEntityCoords(PlayerPedId(), false);
                             TriggerServerEvent(Events.Server.BroadcastHuntedZone, new { Position = position });
@@ -473,10 +686,12 @@ namespace SurviveTheHuntClient
                 {
                     Events.Client.NotifyAboutHuntedZone.EventName(), new Action<dynamic>(data =>
                     {
-                        string playerName = data.PlayerName;
+                        int playerServerId = int.Parse(data.PlayerServerId);
+                        Player player = new Player(GetPlayerFromServerId(playerServerId));
+                        string playerName = player.Name;
                         float nextNotificationTimeout = data.NextNotification;
                         GameState.Hunt.NextMugshotTime = Utility.CurrentTime + TimeSpan.FromSeconds(nextNotificationTimeout);
-                        HuntUI.NotifyAboutHuntedZone(Players[playerName], data.Position, ref GameState);
+                        HuntUI.NotifyAboutHuntedZone(player, data.Position, ref GameState);
                     })
                 },
                 {
@@ -501,6 +716,13 @@ namespace SurviveTheHuntClient
                 Constants.WeaponLoadouts[Teams.Team.Hunters] = deserialized.HuntersWeapons;
                 Constants.WeaponLoadouts[Teams.Team.Hunted] = deserialized.HuntedWeapons;
                 Constants.Vehicles = deserialized.VehicleWhitelist.Vehicles.Select((vehicleName) => (VehicleHash)GetHashKey(vehicleName)).ToArray();
+            
+                // In the event the player was already given weapons, remove them so that the new loadout can be applied.
+                if(Player.Local?.Character != null && PlayerState?.WeaponsGiven == true)
+                {
+                    Ped localPlayerPed = Player.Local.Character;
+                    PlayerState.TakeAwayWeapons(ref localPlayerPed);
+                }
             });
 
             EventHandlers[Events.Client.MarkPlayerDeath] += new Action<float, float, float, Teams.Team>((deathPosX, deathPosY, deathPosZ, team) =>